==> 01.Setting-up-fresh-neo4j-database.ipynb <==
# Initialising out new Neo4j database
Getting started can we connect to our database and see some nodes etc? This notebook uses the native "bolt" driver produced by neo4j themselves to access and run commands.

# We need the connection driver imported
from neo4j.v1 import GraphDatabase

user = "myusername"
password = "mypassword"
connection_path = "bolt://10.0.0.1:7687" # Make sure to set the correct address and port
driver = GraphDatabase.driver(connection_path, auth=(user, password))

Now we can open a session connection to the database

session = driver.session()

Now we can use the session to run a CYPHER query. We will count all the nodes and return it to a result object that we will then print.

result = session.run("MATCH (n) RETURN COUNT(n);")
print(result.data())

Great, there is nothing there since it's a new database but we need to remember to close the session

session.close()

# This is an example query that deletes all nodes in the database 
# DO NOT UNCOMMENT unless you know what you are doing!
# session.run("MATCH (n) DETACH DELETE n;")

# Setting up some constraints

We know the design of the database that we intend to have, e.g. the node types and what will define if they are unique. Hence we can set up some contraints that will also inform the indexing of the database. We know we will have the following:

- *Company* nodes with unique **uid**
- *Country* nodes with unique **codes**
- *Person* nodes with unique **uid**
- *ControllingEntity* with unique **type**
- *Postcode* with unique **uid**
- *Statement* with unique **type**
- *OtherEntity* with unique **uid**

session = driver.session()

session.run("CREATE CONSTRAINT ON (n:Company) ASSERT n.uid IS UNIQUE;")
session.run("CREATE CONSTRAINT ON (n:Person) ASSERT n.uid IS UNIQUE;")
session.run("CREATE CONSTRAINT ON (n:Postcode) ASSERT n.uid IS UNIQUE;")
session.run("CREATE CONSTRAINT ON (n:Other_entity) ASSERT n.uid IS UNIQUE;")
session.run("CREATE CONSTRAINT ON (n:Country) ASSERT n.code IS UNIQUE;")
session.run("CREATE CONSTRAINT ON (n:ControllingEntity) ASSERT n.type IS UNIQUE;")
session.run("CREATE CONSTRAINT ON (n:Statement) ASSERT n.type IS UNIQUE;")

session.close()
