==> 02.Country-Node-creation.ipynb <==
## Storing the countries in Neo4j

We have a master list of ISO countries that we can use to create our base country nodes.

import json
import pandas as pd

iso_data = pd.read_json('./data/iso3166_country_codes.json')

iso_data.head()

### Now we can loop over the data and insert each country with it's corresponding code and name

from neo4j.v1 import GraphDatabase
driver = GraphDatabase.driver("bolt://10.0.0.1:7687", auth=("myusername", "mypassword"))

First we create a list of all the new nodes we wish to create and their associated properties. Then we can connect to the neo4j database and run a Cypher command to insert and create new nodes; we use MERGE in case we have any duplicate entries.

countryList = [{'code': row.iso3166_code, 'name': row['name'].strip().upper()} for ind, row in iso_data.iterrows()]

with driver.session() as session:
    session.run(("UNWIND {list} AS d "
                 "MERGE (c:Country {code: d.code, name: d.name})"),
                {"list": countryList})

## Combining the ISO country codes with secrecy data
We want to incorporate finanical secrecy data into our modelling of the network. We used an edited version of the financial secrecy classification of countries but a similar data set is available to the public as an Excel file here:
- Financial Secrecy Index 2018 Results; https://www.financialsecrecyindex.com/introduction/fsi-2018-results

Be warned the data will take a little bit of cleaning up afterwards it can be used in a similar fashion as indicated below.

secrecy_csv = './data/financialSecrecy.csv'

code2name = {row['code']: row['name'] for row in countryList}

### Country & Nationality mapping

In many instances we will need to be able to convert between country and nationality when connecting people and companies to countries.

We have three different source files that we can use, natively stored as JSON we will convert them to pickles for reasy use within python. The 3 files are:

- Country names to country code mappings
- Country code to country name mappings
- Nationality to country code mappings

def convert_score(datum):
    components = datum.split('-')
    if len(components) == 1:
        return float(components[0])
    else:
        return (float(components[0][1:])+float(components[1][:-1]))/2

secrecy_df = pd.read_csv(secrecy_csv)
secrecy_df['final_secrecy_score'] = secrecy_df['Secrecy Score'].map(convert_score)

secrecy_df.head()

secrecy_df['COUNTRY_CODE'] = secrecy_df.Jurisdiction.map(lambda x: country_names_2_codes.get(x.strip().upper(), 'missing'))

missing_regions = secrecy_df[secrecy_df.COUNTRY_CODE == 'missing'].Jurisdiction.values

missing_regions

from fuzzywuzzy import fuzz
from fuzzywuzzy import process

for region in missing_regions:
    print(region, process.extract(region, country_names_2_codes.keys(), limit=3))

These are certainly things we are missing and need to be added to our data

- St Lucia == ST. LUCIA
- Brunei Darussalam == BRUNEI
- St Vincent & the Grenadines == ST. VINCENT & GRENADINES
- St Kitts & Nevis == ST. KITTS & NEVIS
- United Arab Emirates (Dubai) == UNITED ARAB EMIRATES
- Macao == MACAU
- Czech Republic == CZECHIA

extra_names = ['St Lucia', 'Brunei Darussalam', 'St Vincent & the Grenadines', 'St Kitts & Nevis', 'United Arab Emirates (Dubai)', 'Macao', 'Czech Republic']
extra_names = [name.upper() for name in extra_names]
name4code = ['ST. LUCIA', 'BRUNEI', 'ST. VINCENT & GRENADINES', 'ST. KITTS & NEVIS', 'UNITED ARAB EMIRATES', 'MACAU', 'CZECHIA']

extra_names2codes = {vals[0]: country_names_2_codes.get(vals[1]) for vals in zip(extra_names, name4code)}

country_names_2_codes.update(extra_names2codes)

pd.to_pickle(country_names_2_codes, "data/combined_country_map.pkl")

If we then re-run the above to get the new secrecy numbers we get a new dataframe that we can use to enrich the country nodes.

filtered_df = secrecy_df[secrecy_df.COUNTRY_CODE != 'missing']

with driver.session() as session:
    for i, row in filtered_df.iterrows():
        print(row['COUNTRY_CODE'], row['final_secrecy_score'])
        result = session.run("MATCH (c:Country {code: {code}}) SET c.secrecy_score = {score}", 
                             code=row['COUNTRY_CODE'], 
                             score=row['final_secrecy_score'])

### Country codes to names

country_codes_2_names = json.load(open('data/clean_country_code_map.json', 'r'))

From our data we created clean country codes, do they match with what ISO data we have?

for key in country_codes_2_names.keys():
    result = code2name.get(key, '**')
    if result == '**':
        print(key, country_codes_2_names.get(key))

It looks like we are missing Kosovo which has code XK. DQ seems to be a mistake as it is Dominica and has real code DM. So let's add Kosovo and then save the output to a pkl.

code2name['XK'] = 'KOSOVO'

pd.to_pickle(code2name, "data/clean_country_code_map.pkl")

Let's also fix the DQ in the country_codes_2_names file ...

country_codes_2_names.pop('DQ', None)
country_codes_2_names['DM'] = code2name.get('DM')
print(country_codes_2_names['DM'])

### Country names to codes

We also need a reverse lookup and since there are a variety of ways of doing the lookup let's integrate all the names we have to their respective codes.

country_names_2_codes = json.load(open("data/country_name_2_code_map.json", "r"))

country_names_2_codes = {str(k).upper(): str(v).upper()for k,v in country_names_2_codes.items()}
country_names_2_codes.get('DQ')
country_names_2_codes['DQ'] = 'DM'

print(len(country_codes_2_names))
country_codes_2_names.update({v: k for k,v in code2name.items()})
print(len(country_codes_2_names))
country_codes_2_names.update({v: k for k,v in country_codes_2_names.items()})
print(len(country_codes_2_names))

pd.to_pickle(country_names_2_codes, "data/combined_country_map.pkl")

### Nationality

nationality_2_codes = json.load(open("data/nationality_map.json", "r"))

nationality_2_codes.pop('(blank)', None)
nationality_2_codes['domincan'] = 'DM'

pd.to_pickle(nationality_2_codes, "data/nation_map.pkl")
