==> 04.Inserting-human-PSCs.ipynb <==
# Inserting Individual PSCs into the graph

The largest class of PSC are the individual "human" owners. Here we load in the full dataset and select out the individuals before converting the raw data into a form that will be stored within the neo4j graph.

import pandas as pd
import json
from pandas.io.json import json_normalize
import numpy as np
from neo4j.v1 import GraphDatabase

original_psc_data = pd.read_json('./data/psc_snapshot-2017-09-08.json')
all_records_psc = pd.concat([original_psc_data['company_number'],json_normalize(original_psc_data['data'])],axis=1)
del original_psc_data

all_records_psc.head()

all_records_psc.kind.value_counts()

## Filter the data to only handle individual people with control

human_psc = all_records_psc[all_records_psc.kind == "individual-person-with-significant-control"]
del all_records_psc

human_psc[human_psc['company_number'] == '10262086']

def convert_control_list(control_list):
    """Function to convert a list of controls into a dictionary with the controls as a key and a value of 1"""
    if isinstance(control_list, list):
        return {key: 1 for key in control_list}
    else:
        return {}

human_psc['DICTIONARY_OF_CONTROLS'] = human_psc['natures_of_control'].map(convert_control_list)

human_psc.head()

number_missing_companyID = human_psc[human_psc['company_number'].isnull()].shape[0]
print("WARNING: Missing company number for {} companies in list".format(number_missing_companyID))
print("Removing these rows from data; if significant you should investigate")
human_psc = human_psc[human_psc['company_number'].notnull()]

country_code_map = pd.read_pickle('./data/clean_country_code_map.pkl')
combined_map = pd.read_pickle('./data/combined_country_map.pkl')
nationality_map = pd.read_pickle('./data/nation_map.pkl')

### Defining specific functions to handle human PSCs

These functions will convert the raw data for human PSCs into a format that can then be inderted into Neo4j in an easy fashion.

def psc_dob(record):
    "Function to create dob string identifier for person"
    try:
        dob_segment = "{}/{}".format(str(int(record['date_of_birth.month'])).zfill(2), int(record['date_of_birth.year']))
        return dob_segment
    except TypeError as e:
        return '00/0000'
    except ValueError as e:
        return '00/0000'

def psc_name_id(record):
    "Function to create name string identifier for person"
    try:
        name_segment = "{}_{}_{}".format(str(record['name_elements.surname']).upper(), 
                                         str(record['name_elements.middle_name']).lower(),
                                         str(record['name_elements.forename']).lower())
        return name_segment.replace(' ', '-').replace('nan', '')
    except TypeError as e:
        return 'NONAME__BLANK'
    except AttributeError:
        return 'NONAME__BLANK'
    
def psc_uid(record):
    "Function to create a unique ID for someone from their name and dob"
    components = [psc_name_id(record), psc_dob(record)]
    return {'uid': ':'.join(components)}

def psc_name(record):
    "Function to create name record of person"
    try:
        name_record = {
            'surname': str(record['name_elements.surname']),
            'middle_name': str(record['name_elements.middle_name']),
            'forename': str(record['name_elements.forename']),
            'title': str(record['name_elements.title']),
            'name': record.get('name', '').upper()
        }
        return {k: v.replace('nan', '') for k,v in name_record.items()}
    except TypeError as e:
        return {}
    except AttributeError:
        return {}

def psc_address(record):
    "Function to create a address details for person"
    data = record
    try:
        new_address = {
            'premises': str(data.get('address.premises', '')).replace('nan', ''),
            'address_Line1': str(data.get('address.address_line_1', '')).replace('nan', ''),
            'address_Line2': str(data.get('address.address_line_2', '')).replace('nan', ''),
            'address_PostTown': str(data.get('address.locality', '').replace('nan', '')).upper(),
            'address_POBox': str(data.get('address.po_box', '')).replace('nan', ''),
            'address_County': str(data.get('address.region', '')).replace('nan', '').upper(),
            'address_PostCode': str(data.get('address.postal_code', '')).replace('nan', '').upper(),
            'address_Country': str(data.get('address.country', '')).replace('nan', '').upper(),
            'address_CareOf': str(data.get('address.care_of', '')).replace('nan', '')
        }
        return new_address
    except TypeError as e:
        return {'address_Country': "UNKNOWN"}
    except AttributeError:
        return {'address_Country': "UNKNOWN"}
    
def psc_details(record):
    "Function to create a relationship details for person"
    details = {}
    details['company_id'] = record['links.self'].split('/')[2]
    details['control_kind'] = record['kind']
    details['nationality'] = record['nationality']
    details['ceased_on'] = record['ceased_on']
    dob = psc_dob(record)
    if dob != '00/0000':
        details['DOB'] = dob
    return details
    

def new_record(record):
    "Function to create a new record that is flat"
    uid = psc_uid(record)
    address = psc_address(record)
    return {**uid, **address, 
            **psc_details(record), 
            **psc_name(record), 
            'natures_of_control': record['DICTIONARY_OF_CONTROLS']}


human_psc.iloc[0:5].apply(lambda s: pd.Series(new_record(s)), axis=1)

neo_records_df = human_psc.head(10).apply(lambda s: pd.Series(new_record(s)), axis=1)

neo_records_df['Citizen_of'] = neo_records_df.nationality.apply(lambda x: nationality_map.get(x.lower(), ''))
neo_records_df['Registered_in'] = neo_records_df.address_Country.apply(lambda x: combined_map.get(x.upper(), ''))

neo_records_df.tail(5)

human_psc.ceased_on.value_counts().sum()

## Let us first insert old PSCs that have now ceased controlling a company

If we we have "ceased_on" date then we know we have a former PSC.

neo_records_df = human_psc[human_psc.ceased_on.notnull()].apply(lambda s: pd.Series(new_record(s)), axis=1)

neo_records_df['Citizen_of'] = neo_records_df.nationality.apply(lambda x: nationality_map.get(x.lower(), ''))
neo_records_df['Registered_in'] = neo_records_df.address_Country.apply(lambda x: combined_map.get(x.upper(), ''))

neo_records_df.tail(5)
print(neo_records_df.shape)

#### Need a little more enrichment for when the control relationship ceased

def extend_control_relations(row):
    original = row['natures_of_control']
    original.update({'ceased_on': row['ceased_on']})
    return original

neo_records_df['natures_of_control'] = neo_records_df.apply(lambda x: extend_control_relations(x), axis=1)

We need to convert the dataframe into a list of record details before we start to insert into the Neo4j database

input_data = [v for k,v in neo_records_df.T.to_dict().items()]

driver = GraphDatabase.driver("bolt://10.0.0.1:7687", auth=("myusername", "mypassword"))

**Now** we are ready to run over the data and insert the people and relationships

with driver.session() as session:
    session.run(("UNWIND {list} AS d "
                 "MERGE (c:Person {uid: d.uid}) "
                 "ON CREATE SET c:PSC, c.forename=d.forename, "
                 "c.middle_name=d.middle_name, "
                 "c.surname=d.surname, "
                 "c.title=d.title, "
                 "c.name=d.name, "
                 "c.dob=d.DOB, "
                 "c.nationality=d.nationality, "
                 "c.address_premises=d.premises, "
                 "c.address_Line1=d.address_Line1, "
                 "c.address_Line2=d.address_Line2, "
                 "c.address_PostTown=d.address_PostTown, "
                 "c.address_POBox=d.address_POBox, "
                 "c.address_County=d.address_County, "
                 "c.address_Postcode=d.address_PostCode, "                 
                 "c.address_Country=d.address_Country, "                 
                 "c.uri=d.URI;"), {"list": input_data})

with driver.session() as session:
    session.run(("UNWIND {list} AS d "
                 "MATCH (c:Person {uid: d.uid}) "
                 "MERGE (country:Country {code: d.Registered_in}) "
                 "MERGE (c)-[:REGISTERED_IN]->(country);"), {"list": input_data})

with driver.session() as session:
    session.run(("UNWIND {list} AS d "
                 "MATCH (c:Person {uid: d.uid}) "
                 "MERGE (country:Country {code: d.Citizen_of}) "
                 "MERGE (c)-[:CITIZEN_OF]->(country);"), {"list": input_data})

with driver.session() as session:
    session.run(("UNWIND {list} AS d "
                 "MATCH (p:Person {uid: d.uid}) "
                 "MERGE (ce:ControllingEntity {type: d.control_kind}) "
                 "MERGE (p)-[:HAS_CONTROL_KIND]->(ce);"), {"list": input_data})

with driver.session() as session:
    session.run(("UNWIND {list} AS d "
                 "MATCH (p:Person {uid: d.uid}) "
                 "MERGE (c:Company {uid: d.company_id}) "
                 "MERGE (p)-[r:CONTROLLED]->(c) ON CREATE SET r=d.natures_of_control;"), {"list": input_data})

with driver.session() as session:
    session.run(("UNWIND {list} AS d "
                 "MATCH (p:Person {uid: d.uid}) "
                 "MERGE (pc:Postcode {uid: d.address_PostCode}) "
                 "MERGE (p)-[:REGISTERED_IN]->(pc);"), {"list": input_data})

Let's check how many people, controlled relationships and companies we've connected together

with driver.session() as session:
    result = session.run("MATCH (p:Person)-[r:CONTROLLED]-(c:Company) RETURN COUNT(distinct p), COUNT(distinct r), COUNT(distinct c);")
    print(result.data())

## Now for active PSCs

Let's turn our CYPHER writes into a function we can use in a loop to move over the data. 

Now we can loop over the larger set of data and insert all of the active PSCs

def insert_individual_PSC_nodes_and_relationships(input_data, driver):
    """From a list of dictionaries construct the appropriate nodes and relationships to be inserted into Neo4j
    :param input_data - a list of dictionaries that have all the required information for nodes and relationships
    :param driver - an active driver object to connect to a neo4j instance
    :return """
    
    # Make sure nodes exist for Person
    with driver.session() as session:
        session.run(("UNWIND {list} AS d "
                     "MERGE (c:Person {uid: d.uid}) "
                     "ON CREATE SET c.forename=d.forename, "
                     "c.middle_name=d.middle_name, "
                     "c.surname=d.surname, "
                     "c.title=d.title, "
                     "c.name=d.name, "
                     "c.dob=d.DOB, "
                     "c.nationality=d.nationality, "
                     "c.address_premises=d.premises, "
                     "c.address_Line1=d.address_Line1, "
                     "c.address_Line2=d.address_Line2, "
                     "c.address_PostTown=d.address_PostTown, "
                     "c.address_POBox=d.address_POBox, "
                     "c.address_County=d.address_County, "
                     "c.address_PostCode=d.address_Postcode, "                 
                     "c.address_Country=d.address_Country, "                 
                     "c.uri=d.URI;"), {"list": input_data})
        
    # Connect Person with Country of registration
    with driver.session() as session:
        session.run(("UNWIND {list} AS d "
                     "MATCH (c:Person {uid: d.uid}) "
                     "MERGE (country:Country {code: d.Registered_in}) "
                     "MERGE (c)-[:REGISTERED_IN]->(country);"), {"list": input_data})
    
    # Connect Person with Country of citizenship
    with driver.session() as session:
        session.run(("UNWIND {list} AS d "
                     "MATCH (c:Person {uid: d.uid}) "
                     "MERGE (country:Country {code: d.Citizen_of}) "
                     "MERGE (c)-[:CITIZEN_OF]->(country);"), {"list": input_data})
        
    # Connect person to control_entity
    with driver.session() as session:
        session.run(("UNWIND {list} AS d "
                     "MATCH (p:Person {uid: d.uid}) "
                     "MERGE (ce:ControllingEntity {type: d.control_kind}) "
                     "MERGE (p)-[:HAS_CONTROL_KIND]->(ce);"), {"list": input_data})
        
    # Connect Person to Company it controls
    with driver.session() as session:
        session.run(("UNWIND {list} AS d "
                     "MATCH (p:Person {uid: d.uid}) "
                     "MERGE (c:Company {uid: d.company_id}) "
                     "MERGE (p)-[r:CONTROLS]->(c) ON CREATE SET r=d.natures_of_control;"), {"list": input_data})
        
    # Connect Person to Postcode it controls    
    with driver.session() as session:
        session.run(("UNWIND {list} AS d "
                     "MATCH (p:Person {uid: d.uid}) "
                     "MERGE (pc:Postcode {uid: d.address_PostCode}) "
                     "MERGE (p)-[:REGISTERED_IN]->(pc);"), {"list": input_data})

Using the odo library within blaze we can loop over our input data in chunks

import blaze as bz

proc_records = 0
total_records = human_psc.shape[0] - 59144
for chunk in bz.odo(human_psc[human_psc.ceased_on.isnull()], target=bz.chunks(pd.DataFrame), chunksize=100000):
    neo_records_df = chunk.apply(lambda s: pd.Series(new_record(s)), axis=1)
    neo_records_df['Citizen_of'] = neo_records_df.nationality.apply(lambda x: nationality_map.get(x.lower(), 'UNKNOWN'))
    neo_records_df['Registered_in'] = neo_records_df.address_Country.apply(lambda x: combined_map.get(x.upper(), 'UNKNOWN'))

    input_data = [v for k,v in neo_records_df.T.to_dict().items()]
    del neo_records_df
    
    insert_individual_PSC_nodes_and_relationships(input_data, driver)
    proc_records += len(input_data)
    print("Processed {} of {} .... {:5.2f}% complete".format(proc_records, total_records, 100*proc_records/total_records))
