==> 05.Inserting-Corporate-PSCs.ipynb <==
# Inserting Corporate PSCs

A significant number of controlling entities are the corporations and companies that control other companies. Here we load the data and prepare it for insertion into the graph.

import pandas as pd
import json
from pandas.io.json import json_normalize
import numpy as np
from neo4j.v1 import GraphDatabase

original_psc_data = pd.read_json('../data/psc_snapshot-2017-09-08.json')
all_records_psc = pd.concat([original_psc_data['company_number'],json_normalize(original_psc_data['data'])],axis=1)
del original_psc_data

all_records_psc.head()

all_records_psc.kind.value_counts()

## Filter the data to only handle corporate control

corporate_psc = all_records_psc[all_records_psc.kind == "corporate-entity-person-with-significant-control"]
del all_records_psc

corporate_psc.head()

def convert_control_list(control_list):
    """Function to convert a list of controls into a dictionary with the controls as a key and a value of 1"""
    if isinstance(control_list, list):
        return {key: 1 for key in control_list}
    else:
        return {}

corporate_psc['DICTIONARY_OF_CONTROLS'] = corporate_psc['natures_of_control'].map(convert_control_list)

## It appears that the company_number does NOT correspond to the company_id for corporate PSCs
** INSTEAD we must extract it from the links.self path
e.g. links.self /company/07804290/persons-with-significant-control/corporate-entity/jc9QvYXKVQMFGOTWsXH-XVt25eU **

for k, v in corporate_psc[corporate_psc['identification.registration_number'].isnull()].iloc[1].items():
    print(k, v)

corporate_psc[corporate_psc.company_number == '07804290']

country_code_map = pd.read_pickle('./data/clean_country_code_map.pkl')
combined_map = pd.read_pickle('./data/combined_country_map.pkl')
nationality_map = pd.read_pickle('./data/nation_map.pkl')

### Defining specific functions to handle human PSCs

These functions will convert the raw data for human PSCs into a format that can then be inderted into Neo4j in an easy fashion.

def corporate_psc_name(record):
    "Function to create name string identifier for corporation"
    data = record['name']
    try:
        return data.upper()
    except TypeError as e:
        return 'NONAME__BLANK'
    except AttributeError:
        return 'NONAME__BLANK'
    
def corporate_psc_uid(record):
    "Function to create a unique ID for someone from their name and dob"
    uid = str(record['identification.registration_number'])
    if uid == 'nan':
        uid = '-'.join(corporate_psc_name(record).split()) 
        
    country = str(record['identification.country_registered'])
    if country == 'nan':
        country = record['address.country']
        if country == 'nan':
            country = 'UNKNOWN'
    if isinstance(country, tuple):
        country = 'UNKNOWN'
        
    return {'uid': uid,
            'registered_country': country,
            'legal_authority': str(record['identification.legal_authority']).replace('nan', 'MISSING'),
            'place_registered': str(record['identification.place_registered']).replace('nan', 'MISSING')}


def corporate_psc_address(record):
    "Function to create a address details for person"
    try:
        new_address = {
            'premises': str(record['address.premises']),
            'address_Line1': str(record['address.address_line_1']),
            'address_Line2': str(record['address.address_line_2']),
            'address_PostTown': str(record['address.locality']).upper(),
            'address_POBox': str(record['address.po_box']),
            'address_County': str(record['address.region']).upper(),
            'address_PostCode': str(record['address.postal_code']).upper(),
            'address_Country': str(record['address.country']).upper(),
            'address_CareOf': str(record['address.care_of'])
        }
        return {k: v.replace('nan', '').replace('NAN', '') for k,v in new_address.items()}
    except TypeError as e:
        return {'address_Country': "UNKNOWN"}
    except AttributeError:
        return {'address_Country': "UNKNOWN"}
    
    
def corporate_psc_details(record):
    "Function to create a relationship details for person"
    details = {}
    details['company_id'] = record['links.self'].split('/')[2]
    details['control_kind'] = record['kind']
    return details
    
    
def new_corporate_record(record):
    "Function to create a new record that is flat"
    uid = corporate_psc_uid(record)
    address = corporate_psc_address(record)
    return {**uid, **address, 
            **corporate_psc_details(record), 
            'name': corporate_psc_name(record), 
            'natures_of_control': record['DICTIONARY_OF_CONTROLS']}

Time to test some of these functions

corporate_psc.iloc[0:5].apply(lambda s: pd.Series(new_corporate_record(s)), axis=1)

neo_records_df = corporate_psc.head(10).apply(lambda s: pd.Series(new_corporate_record(s)), axis=1)
neo_records_df['Registered_in'] = neo_records_df.registered_country.apply(lambda x: combined_map.get(x.upper(), ''))

neo_records_df.tail(5)

## Now to insert the Corporate PSCs

Now we can loop over the larger set of data and insert all of the active PSCs

def write_corporate_psc_to_neo(input_data, driver):
    """Function writes corporate records to Neo4j database
    From a list of dictionaries construct the appropriate nodes and relationships to be inserted into Neo4j
    :param input_data - a list of dictionaries that have all the required information for nodes and relationships
    :param driver - an active driver object to connect to a neo4j instance
    :return """
    
    with driver.session() as session:
        session.run(("UNWIND {list} AS d "
                     "MERGE (c:Company {uid: d.uid}) "
                     "ON CREATE SET c.name=d.name, "
                     "c.address_premises=d.premises, "
                     "c.address_Line1=d.address_Line1, "
                     "c.address_Line2=d.address_Line2, "
                     "c.address_PostTown=d.address_PostTown, "
                     "c.address_POBox=d.address_POBox, "
                     "c.address_County=d.address_County, "
                     "c.address_PostCode=d.address_Postcode, "                 
                     "c.address_Country=d.address_Country, "
                     "c.legal_authority=d.legal_authority, "
                     "c.place_registered=d.place_registered, "
                     "c.registered_country=d.registered_country;"), {"list": input_data})

    with driver.session() as session:
        session.run(("UNWIND {list} AS d "
                     "MATCH (c:Company {uid: d.uid}) "
                     "MERGE (country:Country {code: d.Registered_in}) "
                     "MERGE (c)-[:REGISTERED_IN]->(country);"), {"list": input_data})

    with driver.session() as session:
        session.run(("UNWIND {list} AS d "
                     "MATCH (p:Company {uid: d.uid}) "
                     "MERGE (ce:ControllingEntity {type: d.control_kind}) "
                     "MERGE (p)-[:HAS_CONTROL_KIND]->(ce);"), {"list": input_data})    

    with driver.session() as session:
        session.run(("UNWIND {list} AS d "
                     "MATCH (p:Company {uid: d.uid}) "
                     "MERGE (c:Company {uid: d.company_id}) "
                     "MERGE (p)-[r:CONTROLS]->(c) ON CREATE SET r=d.natures_of_control;"), {"list": input_data})

    with driver.session() as session:
        session.run(("UNWIND {list} AS d "
                     "MATCH (c:Company {uid: d.uid}) "
                     "MERGE (pc:Postcode {uid: d.address_PostCode}) "
                     "MERGE (c)-[:REGISTERED_IN]->(pc);"), {"list": input_data})

Using the odo library within blaze we can loop over our input data in chunks

driver = GraphDatabase.driver("bolt://10.0.0.1:7687", auth=("myusername", "mypassword"))

import blaze as bz

proc_records = 0
total_records = corporate_psc.shape[0]
for chunk in bz.odo(corporate_psc, target=bz.chunks(pd.DataFrame), chunksize=50000):
    neo_records_df = chunk.apply(lambda s: pd.Series(new_corporate_record(s)), axis=1)
    neo_records_df['Registered_in'] = neo_records_df.registered_country.apply(lambda x: combined_map.get(str(x).upper(), 'UNKNOWN'))

    input_data = [v for k,v in neo_records_df.T.to_dict().items()]
    del neo_records_df
    
    write_corporate_psc_to_neo(input_data, driver)
    proc_records += len(input_data)
    print("Processed {} of {} .... {:5.2f}% complete".format(proc_records, total_records, 100*proc_records/total_records))
