==> 07.Inserting-PSC-Super-Secure-Persons.ipynb <==
# Loading the "super-secure" PSC control relationships

Some companies make a statement that the PSC needs to be super-secure. Here we build those control relationships and match the appropriate companies to them.

import pandas as pd
import json
from pandas.io.json import json_normalize;

import blaze as bz

First load the data and do some tidying up so we don't waste RAM

original_psc_data = pd.read_json('../data/psc_snapshot-2017-09-08.json')
all_records_psc = pd.concat([original_psc_data['company_number'],json_normalize(original_psc_data['data'])],axis=1)
del original_psc_data

all_records_psc.kind.value_counts()

So there are 404,603 statements of a PSC not being found for a company. Below we see the breakdown in the types of statement they can make

all_records_psc[all_records_psc.kind == 'super-secure-person-with-significant-control'].dropna(axis=1).head()

## Inserting the Super-Secure Controlling Entity into Neo4j

Adding a super-secure ControllingEnity node

from neo4j.v1 import GraphDatabase
driver = GraphDatabase.driver("bolt://10.0.0.1:7687", auth=("myusername", "mypassword"))

kind = 'super-secure-person-with-significant-control'
with driver.session() as session:
    session.run("CREATE (ce:ControllingEntity {type: {kind}})", kind=kind)

## Now to connect companies to statements

1. First let's define a function that takes a company record and creates the relationship back to the super-secure ControllingEntity. 
2. Then we need to loop over all companies that have declared super-secure PSC and insert those into the neo4j database.

def write_super_secure_control(input_data):
    """Function writes super_secure records to Neo4j database"""
    with driver.session() as session:
        session.run(("UNWIND {list} AS d "
                     "MERGE (c:Company {uid: d.company_id}) "
                     "MERGE (ce:ControllingEntity {type: d.kind}) "
                     "MERGE (c)-[:CONTROLS]->(ce);"), 
                    {"list": input_data})

We don't have too many super-sercure people so we can do this in a single insert

super_secure = all_records_psc[all_records_psc.kind == 'super-secure-person-with-significant-control']
super_secure['company_id'] = super_secure['links.self'].map(lambda x: x.split('/')[2])

input_data = [v for k,v in super_secure[['company_id', 'kind']].T.to_dict().items()]
write_super_secure_control(input_data)

## Considering the "exemptions"

Exemptions behave in a very similar way so we will handle those here too but with a "IS_EXEMPT" relationship

kind = 'exemptions'
with driver.session() as session:
    session.run("CREATE (ce:ControllingEntity {type: {kind}})", kind=kind)

def write_exemption_control(input_data):
    """Function writes super_secure records to Neo4j database"""
    with driver.session() as session:
        session.run(("UNWIND {list} AS d "
                     "MERGE (c:Company {uid: d.company_id}) "
                     "MERGE (ce:ControllingEntity {type: d.kind}) "
                     "MERGE (c)-[:IS_EXEMPT]->(ce);"), 
                    {"list": input_data})

exempt = all_records_psc[all_records_psc.kind == 'exemptions']
exempt['company_id'] = exempt['links.self'].map(lambda x: x.split('/')[2])

input_data = [v for k,v in exempt[['company_id', 'kind']].T.to_dict().items()]
write_exemption_control(input_data)
