==> 08.Inserting-missing-PSC-statements.ipynb <==
# Loading the "missing" PSC statements

Some companies make a statement that the PSC has not yet been identified. Here we build those statement nodes and match the appropriate companies to them.

import pandas as pd
import json
from pandas.io.json import json_normalize;

import blaze as bz

First load the data and do some tidying up so we don't waste RAM

original_psc_data = pd.read_json('../data/psc_snapshot-2017-09-08.json')
all_records_psc = pd.concat([original_psc_data['company_number'],json_normalize(original_psc_data['data'])],axis=1)
del original_psc_data

all_records_psc[~all_records_psc.statement.isnull()].statement.map(lambda x: ',' in x).value_counts()

So there are 404,603 statements of a PSC not being found for a company. Below we see the breakdown in the types of statement they can make

all_records_psc.statement.value_counts()

## Inserting the statements into Neo4j

Assuming that we have all the nodes up front we can build the family of statements that a company can register. In theory we should only do this once when we are initially building the database. Afterwards, we will have clashes or produce duplicate nodes unless we MERGE rather than CREATE.

from neo4j.v1 import GraphDatabase
driver = GraphDatabase.driver("bolt://10.0.0.1:7687", auth=("myusername", "mypassword"))

with driver.session() as session:
    for statement_type in all_records_psc.statement.dropna().unique():
        session.run("CREATE (s:Statement {type: {statement_type}})", statement_type=statement_type)

## Now to connect companies to statements

1. First let's define a function that takes a company record and creates the relationship back to the statement. 
2. Then we need to loop over all companies that have made statements and insert those into the neo4j database.
3. We will chunk over the data so as not to put excess strain on the database.

def write_no_psc_statement(input_data):
    """Function writes company records where there is no PSC statement to Neo4j database"""
    with driver.session() as session:
        session.run(("UNWIND {list} AS d "
                     "MERGE (c:Company {uid: d.company_id}) "
                     "MERGE (s:Statement {type: d.statement}) "
                     "MERGE (c)-[:DECLARED]->(s);"), 
                    {"list": input_data})

We shall loop over the data in chunks of 50k records.

no_statement_df = all_records_psc[all_records_psc.statement.notnull()]
no_statement_df['company_id'] = no_statement_df['links.self'].map(lambda x: x.split('/')[2])

for chunk in bz.odo(no_statement_df[['company_id', 'statement']], target=bz.chunks(pd.DataFrame), chunksize=50000):
    print(chunk.shape)
    input_data = [v for k,v in chunk.T.to_dict().items()]
    write_no_psc_statement(input_data)
