==> 09.Inserting-Officers-People.ipynb <==
# Loading the Officers

We have a data file of officers who have a level of control over the company. We need to map these into the system to connect them to the PSC people we are aware of.

import pandas as pd
import json
from pandas.io.json import json_normalize

import blaze as bz

First load some sample data and do some tidying up so we don't waste RAM.

master_path = './data/'

sample_df = pd.read_csv(master_path+'officers.csv', nrows=10000)

sample_df.columns

sample_df[sample_df.end_date.isnull()].iloc[0]

sample_df[sample_df.type == 'Person'].head()

sample_df[sample_df.title.isnull()]

From this and further exploring the data there are two "types" of officer in the data; people and companies.

We will start with how to map People officers into the database. We will also ignore officers who have an "end date" as they are no longer actively connected to the company.

## People who are officers

First we will define a number of functions that can map the data into the node objects that we want to hold for people who are officers.

def officer_dob(record):
    "Function to create dob string identifier for person"
    record.fillna('')
    data = record['partial_date_of_birth']
    try:
        year, month = data.split('-')
        dob_segment = "{}/{}".format(str(month).zfill(2), year)
        return dob_segment
    except (AttributeError, TypeError, ValueError) as e:
        return '00/0000'

def officer_name_id(record):
    "Function to create name string identifier for person"
    fullname = record['name'].split()
    data = {
        'first_name': fullname[0],
        'last_name': record['last_name'],
        'middle_name': ' '.join(fullname[1:-1])
    }
    try:
        name_segment = "{}_{}_{}".format(data.get('last_name', '').upper(), 
                                         data.get('middle_name', '').lower(),
                                         data.get('first_name', '').lower())
        return name_segment.replace(' ', '-')
    except TypeError as e:
        return 'NONAME__BLANK'
    except AttributeError:
        return 'NONAME__BLANK'
    
def officer_uid(record):
    "Function to create a unique ID for someone from their name and dob"
    components = [officer_name_id(record), officer_dob(record)]
    return {'uid': ':'.join(components)}

def officer_name(record):
    "Function to create name record of person"
    fullname = record['name'].split()
    data = {
        'forename': fullname[0].title(),
        'surname': record['last_name'].title(),
        'middle_name': ' '.join(fullname[1:-1]).title(),
        'title': record['title'].title(),
        'name': ' '.join([record['title'], record['name']])
    }    
    try:
        name_record = {
            'surname': data.get('surname', ''),
            'middle_name': data.get('middle_name', ''),
            'forename': data.get('forename', ''),
            'title': data.get('title', ''),
            'name': data.get('name', '').upper()
        }
        return name_record
    except TypeError as e:
        return {}
    except AttributeError:
        return {}

def officer_address(record):
    "Function to create a address details for person"
    record.fillna('', inplace=True)
    address = record['address.street_address'].split('\\n')
    address_Line1 = address[0]
    address_Line2 = ''
    if len(address)>1:
        address_Line2 = address[1]

    data = {
        'address_line_1': address_Line1,
        'address_line_2': address_Line2,
        'locality': record['address.locality'],
        'region': record['address.region'],
        'postal_code': record['address.postal_code'],
        'country': record['address.country']
    }
    try:
        new_address = {
            'address_Line1': data.get('address_line_1', ''),
            'address_Line2': data.get('address_line_2', ''),
            'address_PostTown': data.get('locality', '').upper(),
            'address_POBox': data.get('po_box', ''),
            'address_County': data.get('region', '').upper(),
            'address_PostCode': data.get('postal_code', '').upper(),
            'address_Country': data.get('country', '').upper(),
            'address_InFull': record['address.in_full']
        }
        return new_address
    except TypeError as e:
        return {'address_Country': "UNKNOWN"}
    except AttributeError:
        return {'address_Country': "UNKNOWN"}
    
def officer_details(record):
    "Function to create a relationship details for person"
    details = {}
    details['company_id'] = str(record['company_number'])
    details['officer_kind'] = record['position'].upper()
    details['nationality'] = record['nationality']
    details['ceased_on'] = record['end_date']
    details['source_url'] = record['source_url']
    details['country_of_residence'] = record['country_of_residence']
    details['person_uid'] = str(record['person_uid'])
    dob = record['partial_date_of_birth']
    try:
        year, month = dob.split('-')
        details['DOB'] = "{}/{}".format(str(month).zfill(2), year)
    except (TypeError, ValueError) as e:
        pass
    return details
    

def new_officer_record(record):
    "Function to create a new record that is flat"
    uid = officer_uid(record)
    address = officer_address(record)
    return {**uid, **address, 
            **officer_details(record), 
            **officer_name(record)}

sample_df[sample_df.type == 'Person'].apply(new_officer_record, axis=1).iloc[2]

We will use the reference objects that we have built before to know how to map countries and nationalities

country_code_map = pd.read_pickle('./data/clean_country_code_map.pkl')
combined_map = pd.read_pickle('./data/combined_country_map.pkl')
nationality_map = pd.read_pickle('./data/nation_map.pkl')

Let's wrap up our processing data into a function that can act on an input dataframe object

def officer_etl(original_df):
    original_df = original_df[original_df.type == 'Person']
    original_df = original_df[original_df.end_date.isnull()]
    processed_df = original_df.apply(lambda s: pd.Series(new_officer_record(s)), axis=1)
    processed_df['Citizen_of'] = processed_df.nationality.apply(lambda x: nationality_map.get(x.lower(), ''))
    processed_df['Registered_in'] = processed_df.address_Country.apply(lambda x: combined_map.get(x.lower(), ''))
    processed_df['Resident_in'] = processed_df.country_of_residence.apply(lambda x: combined_map.get(x.lower(), ''))
    input_data = [v for k,v in processed_df.T.to_dict().items()]
    return input_data

Example test of action a sample of the data

test = officer_etl(sample_df.head(20))
print(len(test))
print(test[0])

## Looping through all of the data and pushing it into Neo4j

from neo4j.v1 import GraphDatabase
driver = GraphDatabase.driver("bolt://10.0.0.1:7687", auth=("myusername", "mypassword"))

num=0
done=0
chunks = pd.read_csv(master_path+'officers.csv', chunksize=100000)
for chunk in chunks:
    input_data = officer_etl(chunk)
    num+=1
    done+=len(input_data)
    print(num)
    if len(input_data) > 0:
        with driver.session() as session:
            session.run(("UNWIND {list} AS d "
                         "MERGE (c:Person {uid: d.uid}) "
                         "ON CREATE SET c.forename=d.forename, "
                         "c.middle_name=d.middle_name, "
                         "c.surname=d.surname, "
                         "c.title=d.title, "
                         "c.fullname=d.name, "
                         "c.dob=d.DOB, "
                         "c.person_uid=d.person_uid, "
                         "c.nationality=d.nationality, "
                         "c.address_Line1=d.address_Line1, "
                         "c.address_Line2=d.address_Line2, "
                         "c.address_PostTown=d.address_PostTown, "
                         "c.address_County=d.address_County, "
                         "c.address_PostCode=d.address_Postcode, "                 
                         "c.address_Country=d.address_Country, "                 
                         "c.uri=d.source_url;"), {"list": input_data})

        with driver.session() as session:
            session.run(("UNWIND {list} AS d "
                         "MATCH (c:Person {uid: d.uid}) "
                         "MERGE (country:Country {code: d.Registered_in}) "
                         "MERGE (c)-[:REGISTERED_IN]->(country);"), {"list": input_data})

        with driver.session() as session:
            session.run(("UNWIND {list} AS d "
                         "MATCH (c:Person {uid: d.uid}) "
                         "MERGE (country:Country {code: d.Citizen_of}) "
                         "MERGE (c)-[:CITIZEN_OF]->(country);"), {"list": input_data})

        with driver.session() as session:
            session.run(("UNWIND {list} AS d "
                         "MATCH (p:Person {uid: d.uid}) "
                         "SET p:Officer;"), {"list": input_data})    

        with driver.session() as session:
            session.run(("UNWIND {list} AS d "
                         "MATCH (p:Person {uid: d.uid}) "
                         "MERGE (c:Company {uid: d.company_id}) "
                         "CREATE (p)-[r:IS_OFFICER_OF {type: d.officer_kind, ceased_on: d.ceased_on}]->(c);"), {"list": input_data})
        print("Finished chunk...{}%".format(done/40386527))
    else:
        print("No records in this chunk")
print("DONE!")
