==> 10.Inserting-Officers-Companies.ipynb <==
# Loading the Officers who are companies

We have a data file of officers who have a level of control over the company. We need to map these into the system to connect them to the PSC companies we are aware of.

**NOTE** This notebook is incomplete as we don't have sufficient data in the datafile to connect companies together due to missing Companies House IDs.

import pandas as pd
import json
from pandas.io.json import json_normalize

import blaze as bz

First load some sample data and do some tidying up so we don't waste RAM.

master_path = './data/'

We will loop over the officers file and extract out the Companies that are acting as officers and which are still active to reduce our file sizes.

num=0
chunks = pd.read_csv(master_path+'officers.csv', chunksize=500000)
company_officers = []
for chunk in chunks:
    num += 1
    valid_entries = chunk[((chunk.end_date.isnull()) & (chunk.type == 'Company'))]
    print(num, valid_entries.shape[0])
    company_officers.append(valid_entries)

company_officers_df = pd.concat(company_officers)

company_officers_df.head()

## PROBLEM!

**The data appears incomplete** as the id column does not appear to correlate with the actual Companies House registration id for the company in question. This data needs to be re-enriched in order for these connections to be put into the neo4j database!
